fix an incompatibility with virtualenv (rhbz#742641)
Previously, we manually fixed up the library search path to be
below %{pypyprefix}, by patching pypy/translator/goal/app_main.py (adding
LIBRARY_INSTALLATION_PATH in patch 3, then using sed to change it to the
correct value in %prep after applying the patch).

This led to issues when using pypy with virtualenv, so now (as suggested in
https://codespeak.net/issue/pypy-dev/issue614 ) we simply install the pypy
binary into %{pypyprefix}, and make /usr/bin/pypy be a symlink to it.  This
allows pypy to find its libraries both when run outside a virtualenv, and
within it.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=742641